requirements.txt
streamlit
psycopg2-binary
python-dotenv
st-annotated-text
requests
beautifulsoup4
google-generativeai